+++ services/trade/internal/logic/getsymbollistadminlogic.go
@@ -30,7 +30,7 @@ func NewGetSymbolListAdminLogic(ctx context.Context, svcCtx *svc.ServiceContext)
 func (l *GetSymbolListAdminLogic) GetSymbolListAdmin(in *trade.GetSymbolListAdminReq) (*trade.GetSymbolListAdminResp, error) {
 	cursor, limit := pageutil.Input(in.Page)
 	list, total, err := l.svcCtx.TradeSymbolModel.FindPage(l.ctx, models.TradeSymbolPageFilter{
-		TenantId:   int64(in.TenantId),
+		TenantId:   in.TenantId,
 		MarketType: int64(in.MarketType),
 		Status:     int64(in.Status),
 		Keyword:    in.Keyword,
